Componentize user language list from home page.
